Update ran_num_guess_game.cpp
This is the full code I came up with. Changed my initial structure to a do... while structure. This seems to be the cleanest way to do it. 

If I wanted to make it easier to read, I would make a function like
str2int(x,y) {
getline(std::cin, x)
y = std::stoi(x)
}

then replace all the get lines and str to int conversions to just that one line function. 

Might do that later if I have time.

Otherwise I think I hit all the points needed for this example. Was kinda hard to figure out what to use as it can all essentially do the same.