Create storing the words from a book using dynamic variables
This one is taking it's toll on me. I tried a few things but didn't like how they were going so posting the skeleton here due to being out of time to work on it currently. I will be able to get it done before the end of the day though.